Add files via upload
You are provided with a large number of Wikipedia comments which have been labeled by
human raters for toxic behavior. The types of toxicity are:
Content
The dataset consists of:
comment_text: The text of the Comment
class: The respective class to which the tweet belongs. There are 5 classes -:
 toxic
 severe_toxic
 obscene
 threat
 insult
identity_hateBusiness Objective:
1. Need to get sentiment analysis of tweets gauge its impact and type
Architecture level analysis:
1. Data transformation/Text processing using R/Python
2. Need to get sentiments Analysis and Emotion mining with some charts like histogram,
Density plot, Barplot, pie-plot etc.
3. Deployment through R Shiny or Flask/ Streamlit